pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "AgiteGen"
version         = "0.0.0"            # overwritten by bump_version.py in CI
description     = "Automated cross-platform scaffold & build tool (Aider + Gemini 2.5 Pro + o3)"
requires-python = ">=3.9"
license         = {text = "MIT"}
keywords        = ["cli","aider","gemini","react-native","flutter","nextjs"]
dependencies    = [
  "rich>=13",
  "typer[all]>=0.12",
  "openai>=1.14",
  "httpx>=0.27",
  "pyyaml>=6",
  "jinja2>=3",
  "chromadb>=0.4.22",
  "tqdm>=4",
]

[project.scripts]
agitegen = "agitegen.cli:app"

[tool.setuptools]
packages = ["agitegen"]
